# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Academic Workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get chat history for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "402":
          description: Insufficient balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientBalanceError"

  # ── Documents ─────────────────────────────────────────────────────────────
...
  /projects/{projectId}/attachments:
    get:
      operationId: listAttachments
      tags: [attachments]
      summary: List attachments for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Attachments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attachment"
    post:
      operationId: uploadAttachment
      tags: [attachments]
      summary: Upload an attachment as base64 JSON
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachmentUpload"
      responses:
        "201":
          description: Uploaded attachment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
  /projects/{projectId}/attachments/{attachmentId}:
    delete:
      operationId: deleteAttachment
      tags: [attachments]
      summary: Delete an attachment
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: attachmentId
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/attachments/{attachmentId}/download:
    get:
      operationId: downloadAttachment
      tags: [attachments]
      summary: Download an attachment file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: number
        - name: attachmentId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: File binary
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
        "404":
          description: Not found

  # ── Activities ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Message:
      type: object
      required: [id, projectId, role, content, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    ChatMode:
      type: string
      enum: [generate, revise, reflect, socratic, quiz, summary]
      description: AI writing assistant mode
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        mode:
          $ref: "#/components/schemas/ChatMode"
          default: revise
        tier:
          type: string
          description: AI tier to use (e.g. "free", "standard", "premium"). Defaults to user's preferred tier.
          example: "standard"

    # ── Document ─────────────────────────────────────────────────────────────────
...
    Attachment:
      type: object
      required: [id, projectId, filename, attachmentType, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        filename:
          type: string
        originalName:
          type: string
        mimeType:
          type: ["string", "null"]
        sizeBytes:
          type: ["number", "null"]
        attachmentType:
          type: string
          enum: [instruction, supplement]
        extractedText:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AttachmentUpload:
      type: object
      required: [filename, base64Content, attachmentType]
      properties:
        filename:
          type: string
        base64Content:
          type: string
        mimeType:
          type: string
        attachmentType:
          type: string
          enum: [instruction, supplement]

    # ── Activity ──────────────────────────────────────────────────────────────
...
    InsufficientBalanceError:
      type: object
      properties:
        error:
          type: string
        balanceCents:
          type: integer
        costCents:
          type: integer
        tierName:
          type: string

    # ── User Profile ────────────────────────────────────────────────────────────
...